Tests - Break out SMRFTestCase to own file.
Change the SMRFTestCase to read the base config, test dir as class
variables so all tests can use it.
